requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
Pillow==10.1.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0
django-modeltranslation==0.18.11